requirements.txt
anyascii==0.2.0
asgiref==3.3.4
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
decorator==5.0.9
Django==3.2.3
django-filter==2.4.0
django-modelcluster==5.1
django-pdb==0.6.2
django-taggit==1.4.0
django-treebeard==4.5.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
essential-generators==1.0
et-xmlfile==1.1.0
html5lib==1.1
idna==2.10
ipython==7.24.0
ipython-genutils==0.2.0
jedi==0.18.0
l18n==2020.6.1
matplotlib-inline==0.1.2
openpyxl==3.0.7
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
telepath==0.1.1
traitlets==5.0.5
urllib3==1.26.5
wagtail==2.13
wcwidth==0.2.5
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
